% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_allen_bw.R, R/theme_allen_minimal.R
\docType{data}
\name{font_an}
\alias{font_an}
\title{Assistant font name R variable aliases}
\format{
length 1 character vector

length 1 character vector
}
\usage{
font_an

font_an
}
\description{
\code{font_an} == "\code{Assistant}"

\code{font_an} == "\code{Assistant}"
}
\keyword{datasets}
