% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pres_minimal.R
\docType{data}
\name{font_an}
\alias{font_an}
\title{Average font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_an
}
\description{
\code{font_an} == "\code{Average}"
}
\keyword{datasets}
