% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading_funs.R
\name{grading}
\alias{grading}
\title{Extract highest and lowest grades from a data frame}
\usage{
grading(..., data = NULL)
}
\arguments{
\item{...}{pass of starts_with() since that will make your life easiers}

\item{data}{is a \code{data.frame} that you need  graded}
}
\value{
is your graded data
}
\description{
\code{grading} is just a package that returns grades for exams from highest to  lowest
}
\details{
\code{grading} requires some exam data
}
\examples{
\dontrun{
library(dplyr)

set.seed(123456)
    students = 26
           student_data = tibble(id = 1:26,
           Students = LETTERS,
           Exam_1 = rnorm(students, mean = 85, sd = 5),
           Exam_2 = rnrom(students, mean = 77, sd = 5),
           Exam_3 = rnorm(students, mean = 70, sd = 5))

 your_graded_data = grading(starts_with("exam"),dat = your_exam_Data)
    }
}
