% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmt-loto.R
\name{pmt_loto}
\alias{pmt_loto}
\title{PMT loto machine}
\usage{
pmt_loto(pmt_picks)
}
\arguments{
\item{pmt_picks}{is vector of numbers from the show}
}
\value{
is your pick
}
\description{
\code{pmt_loto} draws random numbers to guess the lottery ball
}
\details{
\code{grading} requires a vector of numbers
}
