% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_allen_minimal.R
\name{theme_allen_minimal}
\alias{theme_allen_minimal}
\title{A \link{ggplot2} that fits my aesthic preferences}
\usage{
theme_allen_minimal(
  base_family = "Assistant",
  base_size = 12.5,
  leg_pos = NULL,
  guide_in_plot = FALSE
)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{leg_pos}{a character vector that can be set to left or right. Default is NULL}

\item{guid_in_plot}{logical to move legend inside plot. Default is FALSE}
}
\description{
A \link{ggplot2} that fits my aesthic preferences
}
\section{Why Assistant?}{

Well for the most part I think it looks good with my LaTeX font of choice  .
}

\examples{
\dontrun{
library(ggplot2)
library(dplyr)

# seminal scatterplot
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(x="Fuel efficiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 scatterplot example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_allen_minimal()

}
}
