% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pres_minimal.R
\name{theme_pres_minimal}
\alias{theme_pres_minimal}
\title{A \link{ggplot2} theme that fits my aesthtic preferences}
\usage{
theme_pres_minimal(base_family = "Average", base_size = 18)
}
\arguments{
\item{base_family, base_size}{base font family and size}
}
\description{
A \link{ggplot2} theme that fits my aesthtic preferences
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

# seminal scatterplot
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(x="Fuel efficiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 scatterplot example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_ipsum()

# seminal bar chart

update_geom_font_defaults()

count(mpg, class) \%>\%
  ggplot(aes(class, n)) +
  geom_col() +
  geom_text(aes(label=n), nudge_y=3) +
  labs(x="Fuel efficiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 bar chart example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_ipsum(grid="Y") +
  theme(axis.text.y=element_blank())
}
}
