% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{update_my_pkgs}
\alias{update_my_pkgs}
\title{A \link{simple function to update R packages}}
\usage{
update_my_pkgs(
  lib.loc = .libPaths()[1],
  repos = getOption("repos"),
  type = c("binary", "source"),
  with_pak = TRUE,
  ...
)
}
\arguments{
\item{lib.loc}{location of lib paths}

\item{repos}{not sure}

\item{type}{whether to compile from binary or source}

\item{with_pak}{whether to use pak to update the package defaults to TRUE}

\item{...}{additional arguments to be passed to old.packages}
}
\description{
A \link{simple function to update R packages}
}
\section{this just keeps everything up to date and make}{
NA
}

\examples{
\dontrun{
update_my_pkgs(with_pak = TRUE)

}
}
